refactor(entity-currency-fiat): extract buildFallbackFiats to internals, selector to selector.ts
