+++ domain/entity/currency-fiat/src/index.ts
@@ -4,3 +4,4 @@ export * from "./registry";
 export * from "./constants";
 export * from "./types";
 export * from "./slice";
+export * from "./selector";
