+++ domain/entity/currency-fiat/src/slice.ts
@@ -1,14 +1,7 @@
 import { createSlice, type PayloadAction } from "@reduxjs/toolkit";
 import type { FiatCurrency } from "./schema";
-import { getFiatCurrencyByTicker } from "./registry";
-import { OFAC_FIAT_TICKERS, FALLBACK_FIAT_TICKERS } from "./constants";
 import type { SupportedFiatsState } from "./types";
-
-function buildFallbackFiats(): FiatCurrency[] {
-  return FALLBACK_FIAT_TICKERS.filter(t => !OFAC_FIAT_TICKERS.has(t))
-    .map(ticker => getFiatCurrencyByTicker(ticker))
-    .filter((c): c is FiatCurrency => c !== undefined);
-}
+import { buildFallbackFiats } from "./internals";
 
 const initialState: SupportedFiatsState = { fiats: buildFallbackFiats() };
 
@@ -30,12 +23,3 @@ export const supportedFiatsSlice = createSlice({
 });
 
 export const { setFiats } = supportedFiatsSlice.actions;
-
-/**
- * Selects the current runtime-supported fiat currencies from the Redux store.
- */
-export function selectSupportedFiats(state: {
-  supportedFiats: SupportedFiatsState;
-}): FiatCurrency[] {
-  return state.supportedFiats.fiats;
-}
